small bug fixed
감정분석값 결과 TinyInt => float으로 변경, 댓글 수정 시 DateTime이 업데이트 되도록 변경

+++ php-Action/RegisterURL.php
@@ -61,7 +61,7 @@ $createDefaultTable = "
     `DateTime` DATETIME NOT NULL,
     `ProfileImageFileName` VARCHAR(25),
     `CommentIndex` INT(11) NOT NULL AUTO_INCREMENT,
-    `EmotionalAnalysisValue` TINYINT,
+    `EmotionalAnalysisValue` FLOAT,
     PRIMARY KEY(`CommentIndex`)
 )";
