+++ php-Action/URLVerification.php
@@ -48,7 +48,7 @@ else {
       `DateTime` DATETIME NOT NULL,
       `ProfileImageFileName` VARCHAR(25),
       `CommentIndex` INT(11) NOT NULL AUTO_INCREMENT,
-      `EmotionalAnalysisValue` TINYINT,
+      `EmotionalAnalysisValue` FLOAT,
       PRIMARY KEY(`CommentIndex`)
   )";
